🚚 move (basic) to (basic)/_component
