+++ src/app/(basic)/_component/Hero.tsx
