+++ src/app/(basic)/layout.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react'
 import { Header } from '@/src/component/Header/Header'
 import { Footer } from '@/src/component/Footer/Footer'
-import { Hero } from '@/src/app/(basic)/Hero'
+import { Hero } from '@/src/app/(basic)/_component/Hero'
 
 type Props = {
   children: ReactNode
